+++ src/solver.rs
@@ -4,10 +4,10 @@ use petgraph::{
     algo::{condensation, toposort},
     dot::Dot,
     graph::{DiGraph, NodeIndex},
-    visit::{EdgeRef, IntoEdgesDirected},
+    visit::EdgeRef,
 };
 
-use crate::schema::{Constraint, DerivedTypeVariable, FieldLabel, Program};
+use crate::{graph::infer_proc_types, schema::{Constraint, DerivedTypeVariable, FieldLabel, Program}};
 
 pub struct Solver<'a> {
     pub program: &'a Program,
@@ -18,25 +18,12 @@ impl Solver<'_> {
         Solver { program }
     }
     pub fn solve(self: Self) -> () {
-        // find the scc in the callgraph, and iterate in post order
-        let sccs = condensation(self.program.call_graph.clone(), true);
-        let topo_sort = toposort(&sccs, None).unwrap();
-        for ind in topo_sort.iter().rev() {
-            let mut constrains: Vec<&Constraint> = Vec::new();
-            // collect constraints for the scc
-            for proc in sccs.node_weight(*ind).unwrap() {
-                for c in self.program.proc_constraints.get(proc).unwrap() {
-                    constrains.push(c);
-                }
-            }
-            // 1. build the initial graph
-            // 2. saturate the graph
-            // 3. collect the set of interesting vars and run pathexpr on them
-            // 4. create sketches for each function
-        }
+        infer_proc_types(self.program);
     }
 
     // TODO Probably should not do this to the whole program? but for a func at a time
+    /// Infer the sketches for a set of constraints.
+    /// Algorithm E.1 in paper.
     pub fn infer_shapes(self: Self) -> () {
         struct Node {
             dtv: DerivedTypeVariable,