+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### TODO
 
-1. add commandline. OK
-   1. path to print the debug graph. OK
+
 2. implement the initial graph build.
+3. Save memory with [object pool](https://www.google.com/search?q=rust+object+pool)
+
+### 2024-05-28
+
+1. support commandline.
+2. From env var get the path to print the debug graph.